Added support to read the QUIC proj files and then read from the QU_* and QP_* files directly.  When the argument supplied to gpuplume is a .proj file from QUIC, defaults will be picked to set the visualization parameters, but information about buildings, sources, and other simulation parameters will come from the QUIC files directly.
In other words, we can create an environment with QUIC and use it directly.

There is some support that is needed for different emitters (cyl, area, and such), but for know most of these are converted to sphere emitters.

Other parameters for concentration calculations will eventually be read from the QP files.